+++ src/main/java/beforespring/yourfood/app/review/domain/event/ReviewCreatedEvent.java
@@ -1,4 +1,4 @@
-package beforespring.yourfood.app.review.service.event;
+package beforespring.yourfood.app.review.domain.event;
 
 import lombok.Getter;
 import org.springframework.validation.annotation.Validated;
@@ -9,14 +9,12 @@ import java.math.RoundingMode;
 
 @Getter
 @Validated
-public class ReviewUpdatedEvent {
+public class ReviewCreatedEvent {
     private final Long restaurantId;
     private final BigDecimal rating;
-    private final BigDecimal beforeRating;
 
-    public ReviewUpdatedEvent(@NotBlank Long restaurantId, @NotBlank Integer rating, @NotBlank Integer beforeRating) {
+    public ReviewCreatedEvent(@NotBlank Long restaurantId, @NotBlank Integer rating) {
         this.restaurantId = restaurantId;
         this.rating = BigDecimal.valueOf(rating).setScale(5, RoundingMode.HALF_UP);
-        this.beforeRating = BigDecimal.valueOf(beforeRating).setScale(5, RoundingMode.HALF_UP);
     }
 }
